+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `--propagate-preemption` flag to `gantry run`.
+
 ## [v1.6.0](https://github.com/allenai/beaker-gantry/releases/tag/v1.6.0) - 2024-06-14
 
 ### Added